pyproject.toml
[tool.poetry]
name = "new-admin-panel-sprint-3"
version = "0.1.0"
description = ""
authors = ["dosxana"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.9"
elasticsearch = "^8.12.0"
psycopg2 = "^2.9.9"
pydantic = "^2.6.1"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
